+++ src/utils.cpp
@@ -22,9 +22,9 @@ stringstream getFormattedQubitLayer(vector<complex<double>> states)
 	return result;
 }
 
-vector<unsigned int> calculateLayerAlloc(int qubitCount, int nodeCount)
+vector<unsigned long long> calculateLayerAlloc(int qubitCount, int nodeCount)
 {
-	vector<unsigned int> result;
+	vector<unsigned long long> result;
 	size_t layerSize = pow(2, qubitCount);
 
 	if(layerSize / 2 < (size_t)nodeCount) {
